apps/rag/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
qdrant-client==1.7.0
google-generativeai==0.3.1
httpx==0.25.2
tqdm==4.66.1
numpy==1.26.2
huggingface_hub>=0.20.0
pydantic-settings
sentence-transformers>=2.4.0
python-dateutil
google-genai